// apps/web/src/audio/types/audio-events.ts
export enum MicroAudioEvent {
  Click = 'click',
  ClickSecondary = 'clickSecondary',
  ToggleOn = 'toggleOn',
  ToggleOff = 'toggleOff',
  Focus = 'focus',
  InputFocus = 'inputFocus',
  InputBlur = 'inputBlur',
  CheckboxCheck = 'checkboxCheck',
  RadioSelect = 'radioSelect',
  HoverOption = 'hoverOption',
  DropdownToggle = 'dropdownToggle',
  SelectOption = 'selectOption',
  ModalOpen = 'modalOpen',
  ModalClose = 'modalClose',
}

export enum SpriteAudioEvent {
  BadgeEarned = 'badgeEarned',
  LevelUp = 'levelUp',
  StreakMaintained = 'streakMaintained',
  StreakBroken = 'streakBroken',
  AnswerCorrect = 'answerCorrect',
  AnswerIncorrect = 'answerIncorrect',
  VoiceRecordingStart = 'voiceRecordingStart',
  VoiceRecordingStop = 'voiceRecordingStop',
  VoiceResponseReceived = 'voiceResponseReceived',
  ToastSuccess = 'toastSuccess',
  ToastInfo = 'toastInfo',
  ToastWarning = 'toastWarning',
  ActivityStart = 'activityStart',
  ActivityComplete = 'activityComplete',
  LessonStart = 'lessonStart',
  LessonComplete = 'lessonComplete',
  XPGain = 'xpGain',
  ErrorSubtle = 'errorSubtle',
  Congrats = 'congrats',
}